requirements.txt
contourpy==1.3.0
cramjam==2.8.4
cycler==0.12.1
et-xmlfile==1.1.0
fastparquet==2024.5.0
fonttools==4.54.1
fsspec==2024.9.0
greenlet==3.1.1
joblib==1.4.2
kiwisolver==1.4.7
matplotlib==3.9.2
MiniSom==2.3.3
mypy==1.11.2
mypy-extensions==1.0.0
numpy==2.1.1
openpyxl==3.1.5
optype==0.6.1
packaging==24.1
pandas==2.2.3
pandas-stubs==2.2.2.240909
patsy==0.5.6
pillow==10.4.0
plotly==5.24.1
psycopg2==2.9.9
pyarrow==17.0.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
scikit-learn==1.5.2
scipy==1.14.1
scipy-stubs==1.4.1a3
six==1.16.0
sklearn-preprocessing==0.1.0
SQLAlchemy==2.0.35
statsmodels==0.14.3
tabulate==0.9.0
tenacity==9.0.0
threadpoolctl==3.5.0
types-psycopg2==2.9.21.20240819
types-pytz==2024.2.0.20240913
types-tabulate==0.9.0.20240106
typing_extensions==4.12.2
tzdata==2024.1
